# swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Unofficial nhentai API
...
paths:
...
  /posts/search:
    get:
      tags:
        - Posts
      summary: Get all posts for a page
      operationId: getPostsForCategory
      parameters:
        - $ref: '#/components/parameters/Query'
        - $ref: '#/components/parameters/PageNumber'
      responses:
        '200':
          $ref: '#/components/responses/PostListWithPage'
        '400':
          $ref: '#/components/responses/WrongRequest'
        '5XX':
          $ref: '#/components/responses/ServerError'
  /posts/homepage:
    get:
      tags:
        - Posts
      summary: Get all posts in homepage
      operationId: getPostsForHomepage
      parameters:
        - $ref: '#/components/parameters/PageNumber'
      responses:
        '200':
          $ref: '#/components/responses/HomepagePostsList'
        '400':
          $ref: '#/components/responses/WrongRequest'
        '5XX':
          $ref: '#/components/responses/ServerError'
  /posts/pages:
    get:
      tags:
        - Posts
      summary: Get number of pages for a search query
      operationId: getPagesForQuery
      parameters:
        - $ref: '#/components/parameters/Query'
      responses:
        '200':
          description: Returns number of pages
          content:
            application/json:
              schema:
                type: object
                properties:
                  pages:
                    type: integer
                    minimum: 0
        '400':
          $ref: '#/components/responses/WrongRequest'
        '5XX':
          $ref: '#/components/responses/ServerError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: query
      name: key
  responses:
...
    PostListWithPage:
      description: Returns post list with additional datas for pagination
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                type: array
                items:
                  $ref: '#/components/schemas/PostPreview'
              pagination:
                $ref: '#/components/schemas/Pagination'
    HomepagePostsList:
      description: Returns homepage posts
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                type: object
                properties:
                  popular:
                    type: array
                    items:
                      $ref: '#/components/schemas/PostPreview'
                  recent:
                    type: array
                    items:
                      $ref: '#/components/schemas/PostPreview'
              pagination:
                $ref: '#/components/schemas/Pagination'
...
    WrongRequest:
      description: Invalid requests parameters
      content:
        application/json:
          schema:
            type: object
            properties:
              statusCode:
                type: integer
              message:
                type: array
                items:
                  type: string
              error:
                type: string
...
    ServerError:
      description: Unexpected server error
  parameters:
    PageNumber:
      in: query
      name: page
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
...
    Query:
      in: query
      name: q
      description: search
      required: true
      schema:
        type: string
...
  schemas:
...
    PostId:
      type: integer
      minimum: 0
      exclusiveMinimum: true
      maximum: 999999
      example: 177013
    PostPreview:
      type: object
      readOnly: true
      properties:
        name:
          type: string
        thumbnail:
          type: string
          format: uri
        lang:
          type: string
        id:
          $ref: '#/components/schemas/PostId'
...
    Pagination:
      type: object
      readOnly: true
      properties:
        page:
          type: integer
          minimum: 0
security:
  - ApiKeyAuth: []